refactor sequence to update dropdowns
