+++ src/applications/gi/utils/helpers.js
@@ -527,3 +527,23 @@ export const getGIBillHeaderText = (automatedTest = false) => {
     ? 'Learn about and compare your GI Bill benefits at approved schools, employers, and VET TEC providers.'
     : 'Learn about and compare your GI Bill benefits at approved schools and employers.';
 };
+
+export const updateLcFilterDropdowns = (dropdowns, target) => {
+  const updatedFieldIndex = dropdowns.findIndex(dropdown => {
+    return dropdown.label === target.id;
+  });
+
+  const selectedOptionIndex = dropdowns[updatedFieldIndex].options.findIndex(
+    option => option.optionValue === target.value,
+  );
+
+  return dropdowns.map(
+    (dropdown, index) =>
+      index === updatedFieldIndex
+        ? {
+            ...dropdown,
+            current: dropdown.options[selectedOptionIndex],
+          }
+        : dropdown,
+  );
+};
